temporal-io/my-python/MY_AI_PROJECTS/langchain/pyproject.toml
[project]
name = "langchain-temporal-integration"
version = "0.1.0"
description = "Langchain + Temporal.io Integration Projects"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.8.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.10.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "mypy>=1.13.0",
]
